+++ pyrepositoryminer/commands/analyze.py
@@ -1,61 +1,48 @@
-from enum import Enum
-from functools import reduce
-from importlib import import_module
-from inspect import isclass
-from multiprocessing import Pool
+from contextlib import contextmanager
 from pathlib import Path
-from sys import stdin
-from typing import List, Optional
+from typing import TYPE_CHECKING, Iterator, List, Optional
 
-from typer import Abort, Argument, Option, echo
+from typer import Argument, Option
 from typer.models import FileText
 
-from pyrepositoryminer.analyze import InitArgs, initialize, worker
-from pyrepositoryminer.metrics import all_metrics
-from pyrepositoryminer.metrics.diffblob.main import DiffBlobMetric
-from pyrepositoryminer.metrics.diffdir.main import DiffDirMetric
-from pyrepositoryminer.metrics.dir.main import DirMetric
-from pyrepositoryminer.metrics.nativeblob.main import NativeBlobMetric
-from pyrepositoryminer.metrics.nativetree.main import NativeTreeMetric
+from pyrepositoryminer.commands.utils.metric import AvailableMetrics
 
-AvailableMetrics = Enum(  # type: ignore
-    # https://github.com/python/mypy/issues/5317
-    "AvailableMetrics",
-    [(k, k) for k in sorted(all_metrics.keys())],
-)
+if TYPE_CHECKING:
+    from multiprocessing.pool import Pool as tcpool
 
 
-def import_metric(import_str: str):  # type: ignore  # TODO make this a metric abc
-    module_str, _, attrs_str = import_str.partition(":")
-    if not module_str or not attrs_str:
-        echo(f'Import string "{import_str}" must be in format "<module>:<attribute>"')
-        raise Abort()
-    try:
-        module = import_module(module_str)
-    except ImportError as e:
-        if e.name != module_str:
-            raise e from None
-        echo(f'Could not import module "{module_str}"')
-        raise Abort() from e
-    try:
-        instance = reduce(getattr, (module, *attrs_str.split(".")))  # type: ignore
-    except AttributeError as e:
-        print(f'Attribute "{attrs_str}" not found in module "{module_str}"')
-        raise Abort() from e
-    if not isclass(instance):
-        print(f'Instance "{instance}" must be a class')
-        raise Abort()
-    parents = (
-        NativeBlobMetric,
-        DiffBlobMetric,
-        NativeTreeMetric,
-        DirMetric,
-        DiffDirMetric,
+@contextmanager
+def make_pool(
+    workers: int,
+    repository: Path,
+    metrics: List[AvailableMetrics],
+    custom_metrics: List[str],
+) -> Iterator["tcpool"]:
+    from multiprocessing import Pool  # pylint: disable=import-outside-toplevel
+
+    from pyrepositoryminer.analyze import (  # pylint: disable=import-outside-toplevel
+        InitArgs,
+        initialize,
     )
-    if not any(issubclass(instance, parent) for parent in parents):  # type: ignore
-        print(f'Instance "{instance}" must subclass a pyrepositoryminer metric class')
-        raise Abort()
-    return instance
+    from pyrepositoryminer.commands.utils.metric import (  # pylint: disable=import-outside-toplevel
+        import_metric,
+    )
+    from pyrepositoryminer.metrics import (  # pylint: disable=import-outside-toplevel
+        all_metrics,
+    )
+
+    with Pool(
+        max(workers, 1),
+        initialize,
+        (
+            InitArgs(
+                repository,
+                tuple({metric.value for metric in metrics} & all_metrics.keys()),
+                tuple(map(import_metric, set(custom_metrics))),
+            ),
+        ),
+    ) as pool:
+        yield pool
 
 
 def analyze(
@@ -71,22 +58,18 @@ def analyze(
     """Analyze commits of a repository.
 
     Either provide the commit ids to analyze on stdin or as a file argument."""
+    from sys import stdin  # pylint: disable=import-outside-toplevel
+
+    from pyrepositoryminer.analyze import (  # pylint: disable=import-outside-toplevel
+        worker,
+    )
+
     metrics = metrics if metrics else []
     ids = (
         id.strip()
         for id in (commits if commits else stdin)  # pylint: disable=superfluous-parens
     )
-    with Pool(
-        max(workers, 1),
-        initialize,
-        (
-            InitArgs(
-                repository,
-                tuple({metric.value for metric in metrics} & all_metrics.keys()),
-                tuple(map(import_metric, set(custom_metrics))),
-            ),
-        ),
-    ) as pool:
+    with make_pool(workers, repository, metrics, custom_metrics) as pool:
         results = (res for res in pool.imap(worker, ids) if res is not None)
         for result in results:
             print(result)
